Add support for /data/media partition layout
With HTC's (unofficial) HBOOT 2.16 firmware came a /data/media partition layout
which merged the userdata and storage partitions into a single one. This allows
them to emulate storage and allow for mtp while not limiting userdata to only two gigabytes.

In order to support both configurations a few changes are necessary:

* Init must be able to mount partitions by their labels
https://gerrit.omnirom.org/5821/

* MountService must be able to use alternate configurations as determined during boot
https://gerrit.omnirom.org/5831/

* The updater-script cannot specify block devices and must use partition names instead
https://gerrit.omnirom.org/6055/

Thanks goto CaptainThrowback for the updated twrp.fstab and helping get permissions
sorted out for the new setup.
ref: https://github.com/CaptainThrowback/omni_device_jewel_TWRP

Change-Id: Ifefbd2128dbeed40828538d5c27ae56fd1849499
Signed-off-by: Paul Keith <javelinanddart@gmail.com>